changed type to float

+++ src/Methods/MethodsTask2.java
@@ -6,13 +6,13 @@ public class MethodsTask2 {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         System.out.println("Enter a number: ");
-        int number = scanner.nextInt();
+        float number = scanner.nextFloat();
         System.out.println("Enter a power: ");
         int power = scanner.nextInt();
         System.out.println("The result is " + calcPower(number, power));
         scanner.close();
     }
-    static float calcPower(int number, int pow) {
+    static float calcPower(float number, int pow) {
         float result = 1;
         if (pow > 0) {
             for (int i = 1; i <= pow; i++) {
